+++ lahuta/types/mdanalysis.py
@@ -26,15 +26,15 @@ class AtomGroupType(Protocol):
         ...
 
     @property
-    def indices(self) -> NDArray[np.int_]:
+    def indices(self) -> NDArray[np.int32]:
         ...
 
     @property
-    def positions(self) -> NDArray[np.float_]:
+    def positions(self) -> NDArray[np.float32]:
         ...
 
     @positions.setter
-    def positions(self, positions: NDArray[np.float_]) -> None:
+    def positions(self, positions: NDArray[np.float32]) -> None:
         ...
 
     @property
@@ -50,7 +50,7 @@ class AtomGroupType(Protocol):
         ...
 
     @property
-    def ids(self) -> NDArray[np.int_]:
+    def ids(self) -> NDArray[np.int32]:
         ...
 
     @property
@@ -66,7 +66,7 @@ class AtomGroupType(Protocol):
         ...
 
     @property
-    def resids(self) -> NDArray[np.int_]:
+    def resids(self) -> NDArray[np.int32]:
         ...
 
     @property
@@ -74,7 +74,7 @@ class AtomGroupType(Protocol):
         ...
 
     @property
-    def vdw_radii(self) -> NDArray[np.float_]:
+    def vdw_radii(self) -> NDArray[np.float16]:
         ...
 
     def copy(self) -> "AtomGroupType":
@@ -87,7 +87,7 @@ class AtomGroupType(Protocol):
         ...
 
     def __getitem__(
-        self, index: Union[int, slice, NDArray[np.int_]]
+        self, index: Union[int, slice, NDArray[np.int32]]
     ) -> "AtomGroupType":
         ...
